# public/swagger.json
{
  "info": {
    "title": "SHOP : API Documentation",
    "version": "1.0",
    "contact": "admin@spericorn.com",
    "description": "There are the api documentation for the student management system"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http",
    "https"
  ],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "swagger": "2.0",
  "paths": {
...
    "/api/password/reset/{token}": {
      "put": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "Authentication"
        ],
        "summary": "Reset Password",
        "description": "Reset Password",
        "operationId": "ResetPassword",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "description": "token provided by email link thorugh (click the link and copy id form url)",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "token": {
                  "type": "string"
                }
              }
            }
          },
          {
            "in": "body",
            "name": "body",
            "description": "enter your new password",
            "required": "on",
            "schema": {
              "type": "object",
              "properties": {
                "password": {
                  "type": "string"
                },
                "confirmPassword": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "requrestBody": {
          "description": "request body",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "password": {
                    "type": "string"
                  },
                  "confirmPassword": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
...
}